Modified _zxBMPRead_1bit, _zxBMPRead_4bit, _zxBMPRead_8bit, _zxBMPRead_TrueColor, _zxBMPRead_RLE8, and _zxBMPRead_RLE4 to read lines by unsigned int. Modified zxImageFileIdent to accept identifier string as an array of unsigned char. Modified zxsMapCreate and zxsMapDraw to correctly handle internal definition of struct _zxs_map_cell. Debugged zxwPixButtonReadData. Modified zxWindowSetFont to specify the name of the fontset by a constant string. Added const keyword to all XPM icon files. Modified zxGetColor to specify the name of the color by a constant string. Modified zxCharStruct to correctly check valid range of charactors.

+++ src/zxwidget.c
@@ -321,7 +321,7 @@ int zxwEditBoxKeyPress(zxWindow *win, zxwEditBox *eb)
   case XK_Home: case XK_Up:  eb->cursor_str_cur = 0;                                                 break;
   case XK_Down: case XK_End: eb->cursor_str_cur = strlen( eb->label );                               break;
   case XK_Left:              if( eb->cursor_str_cur > 0 ) eb->cursor_str_cur--;                      break;
-  case XK_Right:             if( eb->cursor_str_cur < strlen(eb->label) ) eb->cursor_str_cur++;      break;
+  case XK_Right:             if( (size_t)eb->cursor_str_cur < strlen(eb->label) ) eb->cursor_str_cur++;      break;
   case XK_Insert:            zxwCursorInsToggle( eb );                                               break;
   default: ;
     char str[2];
